autoops-mvp/requirements.txt
flask
docker
google-generativeai
streamlit
pandas
requests
